% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-longnostic.R
\name{add-longnostic}
\alias{add-longnostic}
\alias{add_l_max}
\title{Add longnostic to data}
\usage{
add_l_max(data, id, var)
}
\arguments{
\item{data}{dataframe}

\item{id}{id that uniquely identifies individual}

\item{var}{variable to calculate summary of}
}
\value{
dataframe with longnostic
}
\description{
Adds a longnostic to the original dataframe
}
\examples{
add_l_max(data = wages,
          id = id,
          var = lnw)
}
