% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-k-groups.R
\name{add_k_groups}
\alias{add_k_groups}
\title{Add k groups to help explore ids}
\usage{
add_k_groups(data, id, k)
}
\arguments{
\item{data}{data.frame}

\item{id}{id}

\item{k}{number of random groups}
}
\value{
data.frame with additional column, \code{.rand_id} containing \code{k} groups
}
\description{
Add k groups to help explore ids
}
\examples{
library(tidyr)
library(ggplot2)
library(brolgar)

wages \%>\%
  sample_frac_obs(id = id, size = 0.1) \%>\%
  group_by(id) \%>\%
  nest() \%>\%
  mutate(.rand_id = sample(1:10, nrow(.), replace = TRUE)) \%>\%
  unnest() \%>\%
 ggplot(aes(x = lnw,
            y = exper,
            group = id)) + 
 geom_line() + 
 facet_wrap(~.rand_id)
}
