% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-longnostic.R
\name{add_longnostic}
\alias{add_longnostic}
\alias{add_l_max}
\alias{add_l_diff}
\title{Add a longnostic to a dataframe}
\usage{
add_l_max(data, id, var)

add_l_diff(data, id, var, lag = 1)
}
\arguments{
\item{data}{data.frame}

\item{id}{id variable that uniquely identifies an individual}

\item{var}{variable to calculate the longnostic on}

\item{lag}{the lag to use, default to 1 (used with \code{l_diff})}
}
\description{
Match a longnostic to the id of a dataframe. This saves you creating a
longnostic and then joining it back to a dataframe
}
\examples{
add_l_max(data = wages,
          id = id,
          var = lnw)
add_l_diff(data = wages,
          id = id,
          var = lnw)
add_l_diff(data = wages,
          id = id,
          var = lnw,
          lag = 2)
}
