% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brolgar-package.R
\docType{package}
\name{brolgar-package}
\alias{brolgar}
\alias{brolgar-package}
\title{brolgar: Browse Over Longitudinal Data Graphically and Analytically in R}
\description{
Provides a framework of tools to summarise, visualise, and explore longitudinal data. It builds upon the tidy time series data frames used in the 'tsibble' package, and is designed to integrate within the 'tidyverse', and 'tidyverts' (for time series) ecosystems. The methods implemented include calculating features for understanding longitudinal data, including calculating summary statistics such as quantiles, medians, and numeric ranges, sampling individual series, identifying individual series representative of a group, and extending the facet system in 'ggplot2' to facilitate exploration of samples of data. These methods are fully described in the paper "brolgar: An R package to Browse Over Longitudinal Data Graphically and Analytically in R", Nicholas Tierney, Dianne Cook, Tania Prvan (2020) \href{https://arxiv.org/abs/2012.01619}{arXiv:2012.01619}.
}
\details{
\code{brolgar} stands for: \strong{BR}owse over \strong{L}ongitudinal data \strong{G}raphically
and \strong{A}nalytically in \strong{R}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/njtierney/brolgar}
  \item \url{https://brolgar.njtierney.com/}
  \item Report bugs at \url{https://github.com/njtierney/brolgar/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Tierney \email{nicholas.tierney@gmail.com} (\href{https://orcid.org/0000-0003-1460-8722}{ORCID})

Authors:
\itemize{
  \item Di Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
  \item Tania Prvan \email{tania.prvan@mq.edu.au}
}

Other contributors:
\itemize{
  \item Stuart Lee [contributor]
  \item Earo Wang [contributor]
}

}
\keyword{internal}
