% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brolgar-package.R
\docType{package}
\name{brolgar}
\alias{brolgar}
\title{brolgar}
\description{
\code{brolgar} stands for: \strong{BR}owse over \strong{L}ongitudinal data \strong{G}raphically
and \strong{A}nalytically in \strong{R}.
}
\keyword{internal}
