% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-strata.R
\name{facet_strata}
\alias{facet_strata}
\title{Facet data into n strata}
\usage{
facet_strata(n_strata = 12, nrow = NULL, ncol = NULL,
  scales = "fixed", shrink = TRUE, strip.position = "top")
}
\arguments{
\item{n_strata}{Number of facet panels. Default is 12}

\item{nrow}{Number of rows and columns.}

\item{ncol}{Number of rows and columns.}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{shrink}{If \code{TRUE}, will shrink scales to fit output of
statistics, not raw data. If \code{FALSE}, will be range of raw data
before statistical summary.}

\item{strip.position}{By default, the labels are displayed on the top of
the plot. Using \code{strip.position} it is possible to place the labels on
either of the four sides by setting \code{strip.position = c("top",
  "bottom", "left", "right")}}
}
\value{
a ggplot object
}
\description{
Facet data into n strata
}
\examples{
library(ggplot2)
ggplot(world_heights,
       aes(x = year,
           y = height_cm,
           group = country)) + 
  geom_line() + 
  facet_strata()
}
