% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feat_brolgar}
\alias{feat_brolgar}
\alias{feat_three_num}
\title{Feature: Three number summary}
\usage{
feat_three_num(x, ...)

feat_brolgar(x, ...)
}
\arguments{
\item{x}{A vector to extract features from.}

\item{...}{Further arguments passed to other functions.}
}
\description{
This feature returns the three number summary - min, median, and maximum

This feature returns all features in brolgar.
}
