% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feat_monotonic}
\alias{feat_monotonic}
\title{Feature: Monotonics}
\usage{
feat_monotonic(x, ...)
}
\description{
This feature returns monotonic information - does it always increase,
decrease, or is it unvarying?
}
