% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feat_ranges}
\alias{feat_ranges}
\title{Feature: Ranges}
\usage{
feat_ranges(x, ...)
}
\arguments{
\item{x}{A vector to extract features from.}

\item{...}{Further arguments passed to other functions.}
}
\description{
This feature returns the ranges - the minimum, maximum, range difference,
and interquartile range.
}
