% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feat_spread}
\alias{feat_spread}
\title{Feature: Spread}
\usage{
feat_spread(x, ...)
}
\arguments{
\item{x}{A vector to extract features from.}

\item{...}{Further arguments passed to \code{\link[=b_var]{b_var()}}, \code{\link[=b_sd]{b_sd()}},
\code{\link[=b_mad]{b_mad()}},  and \code{\link[=b_iqr]{b_iqr()}}.}
}
\description{
This feature returns measurements of spread: variance, standard deviation,
median absolute distance, and interquartile range
}
