% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-features.R
\docType{import}
\name{features}
\alias{features}
\alias{reexports}
\alias{features_at}
\alias{features_if}
\alias{features_all}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fablelite}{\code{\link[fablelite]{features}}, \code{\link[fablelite]{features_at}}, \code{\link[fablelite]{features_if}}, \code{\link[fablelite]{features_all}}}
}}

