% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-helpers.R
\name{index_regular}
\alias{index_regular}
\title{Index helper - are there regular intervals between each index?}
\usage{
index_regular(.data, index)
}
\arguments{
\item{.data}{data.frame or tsibble}

\item{index}{proposed index column}
}
\value{
logical. TRUE means it is regular, FALSE means not
}
\description{
Index helper - are there regular intervals between each index?
}
\examples{
index_regular(pisa, year)
}
