% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index-helpers.R
\name{index_summary}
\alias{index_summary}
\title{Index helper - summarise the distance between each index}
\usage{
index_summary(data, index)
}
\arguments{
\item{data}{data.frame}

\item{index}{proposed index column}
}
\value{
summary of difference between each index
}
\description{
Index helper - summarise the distance between each index
}
\examples{
index_summary(pisa, year)
}
