% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_slope.R
\name{key_slope}
\alias{key_slope}
\alias{add_key_slope}
\alias{add_key_slope.default}
\title{Fit linear model for each key}
\usage{
key_slope(.data, formula, ...)

add_key_slope(.data, formula)

add_key_slope.default(.data, formula)
}
\arguments{
\item{.data}{tsibble}

\item{formula}{formula}

\item{...}{extra arguments}
}
\value{
tibble with coefficient information
}
\description{
Using \code{key_slope} you can fit a linear model to each key in the \code{tsibble}.
\code{add_key_slope} adds this slope information back to the data, and returns
the full dimension \code{tsibble}.
}
\examples{
key_slope(heights, height_cm ~ year)

}
