% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognostics.R
\name{l_d1}
\alias{l_d1}
\title{Index of interestingness: first order difference
Need to revisit for missing values}
\usage{
l_d1(df, id, var)
}
\arguments{
\item{df}{data.frame to explore}

\item{id}{vector of ids to define which values belong to which individual}

\item{var}{vector of values for individuals, needs to match the id vector}
}
\description{
Compute the maximum of the first order difference of consecutive values for all individuals in the data
}
\examples{
library(tidyverse)
data(wages)
m <- l_d1(wages, "id", "lnw")

}
