% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys_near.R
\docType{data}
\name{l_funs}
\alias{l_funs}
\alias{l_five_num}
\alias{l_three_num}
\title{A named list of the five number summary}
\format{
An object of class \code{list} of length 5.

An object of class \code{list} of length 3.
}
\usage{
l_five_num

l_three_num
}
\description{
Designed for use with the \code{\link[=keys_near]{keys_near()}} function.
}
\examples{
# Specify your own list of summaries
l_ranges <- list(min = b_min,
                 range_diff = b_range_diff,
                 max = b_max,
                 iqr = b_iqr)

heights \%>\%
  key_slope(formula = height_cm ~ year) \%>\%
  keys_near(key = country,
              var = .slope_year,
              funs = l_ranges)
}
\keyword{datasets}
