% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognostics.R
\name{l_length}
\alias{l_length}
\title{Index of interestingness: n subjects}
\usage{
l_length(df, id, var)
}
\arguments{
\item{df}{data.frame to explore}

\item{id}{vector of ids to define which values belong to which individual}

\item{var}{vector of values for individuals, needs to match the id vector}
}
\description{
Compute the number of observations for each individuals in the data
}
\examples{
library(tidyverse)
data(wages)
ns <- l_length(wages, "id", "lnw")

}
