% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R, R/l_slope.R
\name{l_longnostic}
\alias{l_longnostic}
\alias{l_diff}
\alias{l_slope}
\title{Calculating longnostics - statistics of interest for each individual}
\usage{
l_diff(data, id, var, lag = 1)

l_slope(.data, formula, ...)
}
\arguments{
\item{data}{data.frame to explore}

\item{id}{vector of ids to define which values belong to which individual}

\item{var}{vector of values for individuals, needs to match the id vector}

\item{lag}{the lag to use, default to 1 (used with \code{l_diff})}

\item{formula}{character, a formula representing the slope of interest (used in \code{l_slope})}
}
\value{
dataframe with column id and l_\code{statistic}. For example, \code{l_mean} returns the columns id specified, and \code{l_mean}.
}
\description{
Calculate statistical summaries for each individual, referred to as a
\strong{longnostics}, a portmanteau of \strong{long}itudinal and \strong{cognostic} - a
term coined by Tukey which is itself a portmanteau of "cognitive" and
"diagnostic". These \strong{longnostics} all start with \code{l_}, and calculate
a statistic for each individual in the data, who can be identified with
some \code{id}, with the statistic being calculated for a specific variable.
The current \strong{longnostics} implemented are:
\itemize{
\item \code{l_n_obs()} Number of observations
\item \code{l_diff()} Lagged difference (by default, the first order difference)
\item \code{l_slope()} Slope and intercept (given some linear model formula)
}
}
\examples{
l_diff(wages, id, lnw)
l_diff(wages, id, lnw, lag = 2)
l_n_obs(wages, id)
l_slope(wages, id, lnw~exper)
}
\seealso{
The add_l_\code{statistic} set of functions which add a column for each id to the dataframe.
}
