% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longnostics.R
\name{l_max}
\alias{l_max}
\title{Index of interestingness: max}
\usage{
l_max(data, id, var)
}
\arguments{
\item{data}{data.frame}

\item{id}{id variable that uniquely identifies an individual}

\item{var}{variable to calculate the longnostic on}
}
\description{
Compute the maximum value for all individuals in the data
}
\examples{
data(wages)
l_max(wages, "id", "lnw")

}
