% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longnostics.R
\name{l_mean}
\alias{l_mean}
\title{Index of interestingness: mean}
\usage{
l_mean(data, id, var)
}
\arguments{
\item{data}{data.frame}

\item{id}{id variable that uniquely identifies an individual}

\item{var}{variable to calculate the longnostic on}
}
\description{
Compute the mean for all individuals in the data
}
\examples{
data(wages)
l_mean(wages, "id", "lnw")

}
