% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longnostics.R
\name{l_min}
\alias{l_min}
\title{Index of interestingness: min}
\usage{
l_min(data, id, var)
}
\arguments{
\item{data}{data.frame}

\item{id}{id variable that uniquely identifies an individual}

\item{var}{variable to calculate the longnostic on}
}
\description{
Compute the minimum value for all individuals in the data
}
\examples{
data(wages)
l_min(wages, "id", "lnw")

}
