% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognostics.R
\name{l_n_obs}
\alias{l_n_obs}
\title{Index of interestingness: n subjects}
\usage{
l_n_obs(data, id, var)
}
\arguments{
\item{data}{data.frame to explore}

\item{id}{vector of ids to define which values belong to which individual}

\item{var}{vector of values for individuals, needs to match the id vector}
}
\description{
Compute the number of observations for each individuals in the data
}
\examples{
data(wages)
l_n_obs(wages, id, lnw)

}
