% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognostics.R
\name{l_q1}
\alias{l_q1}
\title{Index of interestingness: first quartile}
\usage{
l_q1(data, id, var)
}
\arguments{
\item{data}{data.frame to explore}

\item{id}{vector of ids to define which values belong to which individual}

\item{var}{vector of values for individuals, needs to match the id vector}
}
\description{
Compute the first quartile value for all individuals in the data
}
\examples{
data(wages)
l_q1(wages, id, lnw)

}
