% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognostics.R
\name{l_q3}
\alias{l_q3}
\title{Index of interestingness: third quartile}
\usage{
l_q3(df, id, var)
}
\arguments{
\item{df}{data.frame to explore}

\item{id}{vector of ids to define which values belong to which individual}

\item{var}{vector of values for individuals, needs to match the id vector}
}
\description{
Compute the third quartile value for all individuals in the data
}
\examples{
library(tidyverse)
data(wages)
m <- l_q3(wages, "id", "lnw")

}
