% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longnostics.R
\name{l_sd}
\alias{l_sd}
\title{Index of interestingness: sd}
\usage{
l_sd(data, id, var)
}
\arguments{
\item{data}{data.frame}

\item{id}{id variable that uniquely identifies an individual}

\item{var}{variable to calculate the longnostic on}
}
\description{
Compute the standard deviation for all individuals in the data
}
\examples{
data(wages)
l_sd(wages, "id", "lnw")

}
