% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longnostics.R
\name{l_slope}
\alias{l_slope}
\title{Index of interestingness: slope}
\usage{
l_slope(data, id, formula)
}
\arguments{
\item{data}{data.frame}

\item{id}{id variable that uniquely identifies an individual}

\item{formula}{character, a formula representing the slope of interest}
}
\description{
Compute the maximum value for all individuals in the data
}
\examples{
data(wages)
l_slope(wages, id, lnw~exper)

}
