% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l-summarise.R
\name{l_summarise}
\alias{l_summarise}
\title{Return keys nearest to given summary statistics.}
\usage{
l_summarise(data, key, var, funs = l_five_num)
}
\arguments{
\item{data}{data.frame}

\item{key}{key, which identifies unique observations.}

\item{var}{variable to summarise}

\item{funs}{named list of functions to summarise by. Default is a given
list of the five number summary, \code{l_five_num}.}
}
\value{
data.frame containing keys closest to a given statistic.
}
\description{
Return keys nearest to given summary statistics.
}
\examples{
wages_ts \%>\%
  l_slope(key = id,
          formula = ln_wages ~ xp) \%>\%
  l_summarise(key = id,
              var = l_slope_xp)
               
# Return observations closest to the five number summary of ln_wages
wages_ts \%>\%
  l_summarise(key = id,
              var = ln_wages)
               
# Specify your own list of summaries
l_ranges <- list(min = b_min,
                 range_diff = b_range_diff,
                 max = b_max,
                 iqr = b_iqr)

wages_ts \%>\%
  l_slope(key = id,
          formula = ln_wages ~ xp) \%>\%
  l_summarise(key = id,
              var = l_slope_xp,
              funs = l_ranges)
}
