% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l-summarise-fivenum.R
\name{l_summarise_fivenum}
\alias{l_summarise_fivenum}
\title{Calculate the five number summary of an id for a selected variable}
\usage{
l_summarise_fivenum(data, id, var)
}
\arguments{
\item{data}{data.frame}

\item{id}{id to select}

\item{var}{variable to summarise}
}
\value{
data.frame containing
}
\description{
Calculate the five number summary of an id for a selected variable
}
\examples{
wages \%>\%
  l_slope(id = id,
          formula = lnw ~ exper) \%>\%
  l_summarise_fivenum(id = id,
                      var = l_slope_exper)
                      
wages \%>\%
  l_summarise_fivenum(id = id,
                      var = lnw)
}
