% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognostics.R
\name{lognostic}
\alias{lognostic}
\title{Calculating lognostics}
\arguments{
\item{data}{data.frame to explore}

\item{var}{vector of values for individuals, needs to match the id vector}

\item{id}{vector of ids to define which values belong to which individual}
}
\description{
Lognostics are Cognitive Diagnostics for longitudinal data
}
