% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-longnostic.R
\name{longnostic}
\alias{longnostic}
\title{Calculate a longnostic}
\usage{
longnostic(data, id, var, statistic, l_name, ...)
}
\arguments{
\item{data}{data.frame}

\item{id}{id variable that uniquely identifies an individual}

\item{var}{variable to calculate the longnostic on}

\item{statistic}{the summary statistic to calculate}

\item{l_name}{name to give the new column}

\item{...}{parameters to pass to statistic}
}
\value{
data.frame of summaries
}
\description{
Calculate a longnostic
}
\examples{
longnostic(wages,
             id = id,
             var = lnw,
             mean,
             l_name = m,
             na.rm = TRUE)
}
