% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_obs.R
\name{n_obs}
\alias{n_obs}
\title{Find number of observations of individuals (keys) in a dataset}
\usage{
n_obs(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{extra arguments}
}
\value{
number of observations - the number of individuals
}
\description{
Here, we are not counting the number of rows in the dataset, but rather
we are counting the number of individual, or keys, in the data
}
