% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/near_middle.R
\name{near_middle}
\alias{near_middle}
\title{Return the middle x percent of values}
\usage{
near_middle(x, middle, within)
}
\arguments{
\item{x}{numeric vector}

\item{middle}{percentage you want to center around}

\item{within}{percentage around center}
}
\value{
logical vector
}
\description{
Return the middle x percent of values
}
\examples{
x <- runif(20)
near_middle(x = x,
            middle = 0.5,
            within = 0.2)
            
library(dplyr)
l_min(wages, id, lnw) \%>\%
  filter(near_middle(l_min, 0.5, 0.1))
            
}
