% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-reflect.R
\name{reflect_palette}
\alias{reflect_palette}
\title{reflective palettes}
\usage{
reflect_palette(palette = "hawaii", pal_dir = "left", n = 5,
  alpha = NULL, begin = 0, end = 1, direction = 1, ...)
}
\arguments{
\item{palette}{character}

\item{pal_dir}{direction of palette reflection}

\item{n}{number of colours}

\item{alpha}{transparency}

\item{begin}{start point}

\item{end}{end point}

\item{direction}{direction}

\item{...}{extra arguments}
}
\value{
palette
}
\description{
reflective palettes
}
\examples{
library(dplyr)
heights_near <- key_slope(heights,
         height_cm ~ year) \%>\%
 keys_near(key = country,
           var = .slope_year) \%>\%
 left_join(heights, by = "country")

library(ggplot2)
ggplot(heights_near,
      aes(x = year,
          y = height_cm,
          group = country,
          colour = stat)) +
 geom_line() +
 scale_colour_reflective(palette = "batlow",
                         pal_dir = "left")

}
