% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_minima}
\alias{safe_minima}
\alias{max_if}
\title{Calculate maximum and return NA if input is empty}
\usage{
max_if(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{additional arguments for max}
}
\value{
either NA or the maximum value
}
\description{
This is a helper function that returns NA if length(x) == 0 (e.g., is
numeric(0)), otherwise calculates the maximum
}
\examples{
\dontrun{
max_if(numeric(0))
}
}
