% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-frac-n-obs.R
\name{sample-n-frac-key}
\alias{sample-n-frac-key}
\alias{sample_n_key}
\alias{sample_frac_key}
\alias{sample_frac_key.data.frame}
\title{Sample a number or fraction of keys to explore}
\usage{
sample_n_key(.data, size, ...)

sample_frac_key(.data, size, ...)

\method{sample_frac_key}{data.frame}(.data, key, size, ...)
}
\arguments{
\item{.data}{data.frame to explore}

\item{size}{The number or fraction of observations, depending on the
function used. In \code{sample_n_key}, it is a number > 0, and in
\code{sample_frac_key} it is a fraction, between 0 and 1.}

\item{...}{extra arguments}

\item{key}{vector of keys to define which values belong to which individual}

\item{key}{vector of the unique identifier}
}
\value{
data.frame with fewer observations of key
}
\description{
Sample a number or fraction of keys to explore
}
\examples{
library(ggplot2)
sample_n_key(wages_ts,
             size = 10) \%>\%
  ggplot(aes(x = exper,
             y = uerate,
             group = id)) + 
  geom_line()
library(ggplot2)
sample_frac_key(wages_ts,
                0.1) \%>\%
  ggplot(aes(x = exper,
             y = uerate,
             group = id)) + 
  geom_line()
}
