% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-frac-n-obs.R
\name{sample_n_obs}
\alias{sample_n_obs}
\alias{sample-n-frac-obs}
\alias{sample_frac_obs}
\title{Sample a fraction of ids to explore}
\usage{
sample_n_obs(data, id, size)

sample_frac_obs(data, id, size)
}
\arguments{
\item{data}{data.frame}

\item{id}{id of interest}

\item{size}{The number or fraction of observations, depending on the
function used. In \code{sample_n_obs}, it is a number > 0, and in
\code{sample_frac_obs} it is a fraction, between 0 and 1.}
}
\value{
data.frame with fewer observations of id
}
\description{
Sample a fraction of ids to explore
}
\examples{
library(ggplot2)
sample_n_obs(wages,
             id,
             10) \%>\%
  ggplot(aes(x = exper,
             y = uerate,
             group = id)) + 
  geom_line()
library(ggplot2)
sample_frac_obs(wages,
                id,
                0.1) \%>\%
  ggplot(aes(x = exper,
             y = uerate,
             group = id)) + 
  geom_line()
}
