% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-frac-n-obs.R
\name{sample_n_key}
\alias{sample_n_key}
\title{Sample a fraction of keys to explore}
\usage{
sample_n_key(.data, size, ...)
}
\arguments{
\item{.data}{data.frame to explore}

\item{size}{The number or fraction of observations, depending on the
function used. In \code{sample_n_obs}, it is a number > 0, and in
\code{sample_frac_obs} it is a fraction, between 0 and 1.}

\item{...}{extra arguments}

\item{key}{vector of keys to define which values belong to which individual}
}
\value{
data.frame with fewer observations of id
}
\description{
Sample a fraction of keys to explore
}
\examples{
library(ggplot2)
sample_n_key(ts_wages,
             size = 10) \%>\%
  ggplot(aes(x = exper,
             y = uerate,
             group = id)) + 
  geom_line()
}
