% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal-reflect.R
\name{scale_colour_reflective}
\alias{scale_colour_reflective}
\title{ggplot2 scale colour reflective}
\usage{
scale_colour_reflective(palette = "hawaii", pal_dir = "left", ...)
}
\arguments{
\item{palette}{character}

\item{pal_dir}{character}

\item{...}{extra arguments}
}
\value{
ggplot
}
\description{
ggplot2 scale colour reflective
}
\examples{
library(dplyr)
heights_near <- key_slope(heights,
         height_cm ~ year) \%>\%
 keys_near(key = country,
           var = .slope_year) \%>\%
 left_join(heights, by = "country")

library(ggplot2)
ggplot(heights_near,
      aes(x = year,
          y = height_cm,
          group = country,
          colour = stat)) +
 geom_line() +
 scale_colour_reflective(palette = "batlow",
                         pal_dir = "left")
}
