% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_near_quant.R
\name{stat-near-quant}
\alias{stat-near-quant}
\alias{l_max_near_quant}
\alias{add_stat_near_quant}
\title{Obtain IDs closest to each quantile for a given stat}
\usage{
l_max_near_quant(data, id, var, probs = c(0.25, 0.5, 0.75),
  nearest_within = .Machine$double.eps^0.5)

add_stat_near_quant(data, id, var, probs = c(0.25, 0.5, 0.75),
  nearest_within = .Machine$double.eps^0.5)
}
\arguments{
\item{data}{data.frame}

\item{id}{id}

\item{var}{variable to calculate stat on}

\item{probs}{vector of quantiles}

\item{nearest_within}{select stats within this this amount (default is
\code{.Machine$double.eps^0.5})}
}
\value{
dataframe with columns \code{id} and \code{nearest_quant}
}
\description{
Obtain IDs closest to each quantile for a given stat
}
\examples{
 l_max_near_quant(data = wages,
                  id = id,
                  var = lnw,
                  nearest_within = 0.01)
 l_max_near_quant(data = wages,
                  id = id,
                  var = lnw,
                  probs = seq(0.1, 0.9, by = 0.1),
                  nearest_within = 0.01)
# dplyr::inner_join(data, l_max_data_near, by = "id")
add_stat_near_quant(wages,
                    id,
                    lnw)
}
