% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify-key.R
\name{stratify_key}
\alias{stratify_key}
\title{Add k groups to help explore ids}
\usage{
stratify_key(.data, n_strata, ...)
}
\arguments{
\item{.data}{data.frame to explore}

\item{n_strata}{number of random groups to create}

\item{...}{extra arguments}
}
\value{
data.frame with additional column, \code{.strata} containing \code{n_strata} groups
}
\description{
Add k groups to help explore ids
}
\examples{
library(ggplot2)
library(brolgar)

ts_wages \%>\%
  sample_frac_key(size = 0.1) \%>\%
  stratify_key(10) \%>\%
 ggplot(aes(x = lnw,
            y = exper,
            group = id)) + 
 geom_line() + 
 facet_wrap(~.strata)
}
