% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wages}
\alias{wages}
\title{Wages data from NLSY}
\description{
This data contains measurements on hourly wages by years in
the workforce, with education and race as covariates.
It comes from J. D. Singer and J. B. Willett.
Applied Longitudinal Data Analysis.
Oxford University Press, Oxford, UK, 2003.
http://gseacademic.harvard.edu/alda/
}
