% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-world-height.R
\docType{data}
\name{world_heights}
\alias{world_heights}
\title{World Height Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1499 rows and 4 columns.}
\usage{
world_heights
}
\description{
Average male heights in 165 countries from 1810-1989, with a
smaller number of countries from 1500-1800.
}
\details{
It contains the variables:
\itemize{
\item code: country code
\item country: the country
\item year: Year
\item height_cm: average height in centimeters for males
}

For more information, see Joerg Baten and Matthias Blum, "Why are
you tall while others are short? Agricultural production and
other proximate determinants of global heights", European
Review of Economic History 18 (2014), 144–165. Data available
from http://hdl.handle.net/10622/IAEKLA, accessed via the Clio
Infra website.
}
\keyword{datasets}
