% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_summaries.R
\name{b_min}
\alias{b_min}
\alias{b_summaries}
\alias{b_max}
\alias{b_median}
\alias{b_q25}
\alias{b_q75}
\alias{b_range}
\alias{b_range_diff}
\title{Brolgar summaries (b_summaries)}
\usage{
b_min(x, na.rm = TRUE, ...)

b_max(x, na.rm = TRUE, ...)

b_median(x, na.rm = TRUE, ...)

b_q25(x, na.rm = TRUE, ...)

b_q75(x, na.rm = TRUE, ...)

b_range(x, na.rm = TRUE, ...)

b_range_diff(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{whether to remove NA values. Default is TRUE}

\item{...}{other arguments to pass}
}
\description{
Customised summaries of vectors with appropriate defaults for longitudinal
data. This includes minimum, maximum, median, q25 and q75. The functions
are prefixed with \code{b_} to assist with autocomplete. The defaults are to use
\code{type = 8} for quantiles, and \code{na.rm = TRUE}.
}
