% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feat_brolgar}
\alias{feat_brolgar}
\title{Feature: All from brolgar}
\usage{
feat_brolgar(x, ...)
}
\arguments{
\item{x}{A vector to extract features from.}

\item{...}{Further arguments passed to other functions.}
}
\description{
This feature returns all features in brolgar.
}
