% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feat_brolgar}
\alias{feat_brolgar}
\title{Feature: All from brolgar}
\usage{
feat_brolgar(x, ...)
}
\description{
This feature returns measurements of spread: variance, standard deviation,
median absolute distance, and interquartile range
}
