% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feat_five_num}
\alias{feat_five_num}
\title{Feature: Five number summary}
\usage{
feat_five_num(x, ...)
}
\arguments{
\item{x}{A vector to extract features from.}

\item{...}{Further arguments passed to other functions.}
}
\description{
This feature returns the five number summary: minimum, q25, median, q75,
and maximum.
}
