% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feat_ranges}
\alias{feat_ranges}
\title{Feature: Ranges}
\usage{
feat_ranges(x, ...)
}
\arguments{
\item{x}{A vector to extract features from.}

\item{...}{Further arguments passed to \code{\link[=b_min]{b_min()}}, \code{\link[=b_max]{b_max()}},
\code{\link[=b_range_diff]{b_range_diff()}},  and \code{\link[=b_iqr]{b_iqr()}},}
}
\description{
This feature returns the ranges - the minimum, maximum, range difference,
and interquartile range.
}
