% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feat_spread}
\alias{feat_spread}
\title{Feature: Spread}
\usage{
feat_spread(x, ...)
}
\arguments{
\item{x}{A vector to extract features from.}

\item{...}{Further arguments passed to other functions.}
}
\description{
This feature returns measurements of spread: variance, standard deviation,
median absolute distance, and interquartile range
}
