% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-n-obs.R
\name{filter_n_obs}
\alias{filter_n_obs}
\alias{filter_n_obs.tbl_ts}
\title{Filter by the number of observations for a \code{key}.}
\usage{
filter_n_obs(.data, filter, ...)

\method{filter_n_obs}{tbl_ts}(.data, filter, ...)
}
\arguments{
\item{.data}{data.frame}

\item{filter}{A description of how you want to filter the number of
observations for each \code{key}, in terms of \code{n_obs}. See examples for more
detail.}
}
\value{
data.frame filtered by the number of observations, with an
additional column \code{n_obs}, which contains the number of observations for
each \code{key}.
}
\description{
When exploring longitudinal data it can be useful to filter by the number of
observations in a compact way. \code{filter_n_obs} allows for the user to
filter by the number of observations for each \code{key}.
}
\examples{
wages_ts \%>\% filter_n_obs(n_obs > 10)
wages_ts \%>\% filter_n_obs(n_obs == 2)

}
