% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-n-obs.R
\name{filter_n_obs}
\alias{filter_n_obs}
\title{Filter by the number of observations for an individual}
\usage{
filter_n_obs(data, id, var, filter)
}
\arguments{
\item{data}{data.frame to explore}

\item{id}{vector of ids to define which values belong to which individual}

\item{var}{vector of values for individuals, needs to match the id vector}

\item{filter}{an expression you want to filter by. For example (\code{l_n_obs > 10}) would filter those observations with greater than 10 observations.}
}
\value{
dataframe filtered by the number of observations
}
\description{
Filter by the number of observations for an individual
}
\examples{
filter_n_obs(wages,
             id = id,
             filter = l_n_obs > 10)

filter_n_obs(wages,
             id = id,
             filter = l_n_obs == 2)

}
