% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys_near.R
\name{keys_near}
\alias{keys_near}
\alias{keys_near.default}
\title{Return keys nearest to a given statistics or summary.}
\usage{
keys_near(.data, ...)

\method{keys_near}{default}(.data, ...)
}
\arguments{
\item{.data}{tsibble}

\item{...}{extra arguments to pass to \code{mutate_at} when performing the summary
as given by \code{funs}.}
}
\value{
data.frame containing keys closest to a given statistic.
}
\description{
Return keys nearest to a given statistics or summary.
}
\examples{
keys_near(heights, height_cm)

}
