% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognostics.R
\name{l_mean}
\alias{l_mean}
\title{Index of interestingness: mean}
\usage{
l_mean(df, id, var)
}
\arguments{
\item{id}{vector of ids to define which values belong to which individual}

\item{d}{vector of values for individuals, needs to match the id vector}
}
\description{
Compute the mean for all individuals in the data
}
\examples{
library(tidyverse)
data(wages)
m <- l_mean(wages, "id", "lnw")

}
