% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longnostics.R
\name{l_median}
\alias{l_median}
\title{Index of interestingness: median}
\usage{
l_median(data, id, var)
}
\arguments{
\item{data}{data.frame}

\item{id}{id variable that uniquely identifies an individual}

\item{var}{variable to calculate the longnostic on}
}
\description{
Compute the median value for all individuals in the data
}
\examples{
data(wages)
l_median(wages, id, lnw)

}
