% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognostics.R
\name{l_min}
\alias{l_min}
\title{Index of interestingness: min}
\usage{
l_min(df, id, var)
}
\arguments{
\item{df}{data.frame to explore}

\item{id}{vector of ids to define which values belong to which individual}

\item{var}{vector of values for individuals, needs to match the id vector}
}
\description{
Compute the minimum value for all individuals in the data
}
\examples{
library(tidyverse)
data(wages)
m <- l_min(wages, "id", "lnw")

}
