% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longnostics.R
\name{l_n_obs}
\alias{l_n_obs}
\title{Index of interestingness: n subjects}
\usage{
l_n_obs(data, id, var)
}
\arguments{
\item{data}{data.frame}

\item{id}{id variable that uniquely identifies an individual}

\item{var}{variable to calculate the longnostic on}
}
\description{
Compute the number of observations for each individuals in the data
}
\examples{
data(wages)
l_n_obs(wages, id, lnw)

}
