% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longnostics.R
\name{l_q1}
\alias{l_q1}
\title{Index of interestingness: first quartile}
\usage{
l_q1(data, id, var)
}
\arguments{
\item{data}{data.frame}

\item{id}{id variable that uniquely identifies an individual}

\item{var}{variable to calculate the longnostic on}
}
\description{
Compute the first quartile value for all individuals in the data
}
\examples{
data(wages)
l_q1(wages, id, lnw)

}
