% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longnostics.R
\name{l_q3}
\alias{l_q3}
\title{Index of interestingness: third quartile}
\usage{
l_q3(data, id, var)
}
\arguments{
\item{data}{data.frame}

\item{id}{id variable that uniquely identifies an individual}

\item{var}{variable to calculate the longnostic on}
}
\description{
Compute the third quartile value for all individuals in the data
}
\examples{
data(wages)
l_q3(wages, id, lnw)

}
