% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognostics.R
\name{l_sd}
\alias{l_sd}
\title{Index of interestingness: sd}
\usage{
l_sd(d, id)
}
\arguments{
\item{d}{vector of values for individuals, needs to match the id vector}

\item{id}{vector of ids to define which values belong to which individual}
}
\description{
Compute the standard deviation for all individuals in the data
}
\examples{
library(tidyverse)
df <- tibble(d=c(1,2,3,1,5,10,12), id=c(rep("A", 3), rep("B", 4)))
l_sd(df$d, df$id)

}
