% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognostics.R
\name{l_sd}
\alias{l_sd}
\title{Index of interestingness: sd}
\usage{
l_sd(df, id, var)
}
\arguments{
\item{df}{data.frame to explore}

\item{id}{vector of ids to define which values belong to which individual}

\item{var}{vector of values for individuals, needs to match the id vector}
}
\description{
Compute the standard deviation for all individuals in the data
}
\examples{
library(tidyverse)
data(wages)
s <- l_sd(wages, "id", "lnw")

}
