% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognostics.R
\name{l_slope}
\alias{l_slope}
\title{Index of interestingness: slope}
\usage{
l_slope(df, id, formula)
}
\arguments{
\item{df}{data.frame to explore}

\item{id}{vector of ids to define which values belong to which individual}

\item{formula}{character, a formula representing the slope of interest}
}
\description{
Compute the maximum value for all individuals in the data
}
\examples{
library(tidyverse)
data(wages)
m <- l_slope(wages, "id", "lnw~exper")

}
