% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_slope.R
\name{l_slope}
\alias{l_slope}
\alias{add_l_slope}
\title{Fit linear model for each key}
\usage{
l_slope(.data, formula, ...)

add_l_slope(.data, formula)
}
\arguments{
\item{.data}{data.frame}

\item{formula}{formula}

\item{...}{extra arguments}
}
\value{
data.frame with coefficient information
}
\description{
Using \code{l_slope} you can fit a linear model to each key in the \code{tsibble}.
\code{add_l_slope} adds this slope information back to the data, and returns
the full dimension \code{tsibble}.
}
