% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-key-obs.R
\name{n_obs}
\alias{n_obs}
\alias{n_key_obs}
\alias{n_key_obs.tbl_ts}
\alias{add_n_key_obs}
\alias{add_n_key_obs.tbl_ts}
\title{Find number of observations for each key in a dataset}
\usage{
n_key_obs(.data, ...)

\method{n_key_obs}{tbl_ts}(.data, ...)

add_n_key_obs(.data, ...)

\method{add_n_key_obs}{tbl_ts}(.data, ...)
}
\arguments{
\item{.data}{data.frame}

\item{...}{extra arguments}
}
\value{
number of observations - the number of individuals
}
\description{
Here, we are not counting the number of rows in the dataset, but rather
we are counting the number of individual, or keys, in the data.
}
