% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-obs.R
\name{n_obs}
\alias{n_obs}
\title{Return the number of observations}
\usage{
n_obs(x, names = TRUE)
}
\arguments{
\item{x}{vector or dataframe}

\item{names}{logical; If TRUE the result is a named vector named "n_obs", else
it is just the number of observations.}
}
\value{
number of observations
}
\description{
Returns the number of obserations of a vector or data.frame. It uses
\code{vctrs::vec_size()} under the hood.
}
\examples{
n_obs(iris)
n_obs(1:10)
# return the number of observations
wages \%>\% 
  features(id, n_obs)
}
