% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearests.R
\name{near_between}
\alias{near_between}
\title{Return x percent to y percent of values}
\usage{
near_between(x, from, to)
}
\arguments{
\item{x}{numeric vector}

\item{from}{the lower bound of percentage}

\item{to}{the upper bound of percentage}
}
\value{
logical vector
}
\description{
Return x percent to y percent of values
}
\examples{
x <- runif(20)

near_middle(x = x,
            middle = 0.5,
            within = 0.2)

library(dplyr)
wages_ts \%>\% feaures(ln_wages, list(min = min)) \%>\%
  filter(near_between(min, 0.1, 0.9))
}
