% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/near_quantile.R
\name{near_quantile}
\alias{near_quantile}
\title{Which values are nearest to a (single) quantile}
\usage{
near_quantile(x, probs, tol)
}
\arguments{
\item{x}{vector}

\item{probs}{single quantile to calculate}

\item{tol}{the tolerance in terms of x that you will accept near to the
quantile}
}
\value{
logical vector of TRUE/FALSE
}
\description{
Which values are nearest to a (single) quantile
}
\examples{
x <- runif(20)
near_quantile(x, 0.5, 0.05)

library(dplyr)
l_min(wages,id, lnw) \%>\% 
  mutate(min_near_median = near_quantile(l_min, 0.5, 0.01)) \%>\%
  filter(min_near_median)
l_min(wages,id, lnw) \%>\% 
  mutate(min_near_median = near_quantile(l_min, 0.5, 0.01)) \%>\%
  filter(min_near_median)
}
