% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{features}
\alias{features_at}
\alias{features_if}
\alias{features_all}
\alias{as_tsibble}
\alias{n_keys}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fabletools}{\code{\link[fabletools]{features}}, \code{\link[fabletools:features]{features_all}}, \code{\link[fabletools:features]{features_at}}, \code{\link[fabletools:features]{features_if}}}

  \item{tsibble}{\code{\link[tsibble:as-tsibble]{as_tsibble}}, \code{\link[tsibble:key-data]{n_keys}}}
}}

