% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify-keys.R
\name{stratify_keys}
\alias{stratify_keys}
\title{Stratify the keys into random groups to facilitate exploration}
\usage{
stratify_keys(.data, n_strata, along = NULL, ...)
}
\arguments{
\item{.data}{data.frame to explore}

\item{n_strata}{number of random groups to create}

\item{along}{variable to stratify along}

\item{...}{extra arguments}
}
\value{
data.frame with column, \code{.strata} containing \code{n_strata} groups
}
\description{
Stratify the keys into random groups to facilitate exploration
}
\examples{
library(ggplot2)
library(brolgar)

wages_ts \%>\%
  sample_frac_keys(size = 0.1) \%>\%
  stratify_keys(10) \%>\%
 ggplot(aes(x = ln_wages,
            y = xp,
            group = id)) + 
 geom_line() + 
 facet_wrap(~.strata)
}
