% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-world-height.R
\docType{data}
\name{world_heights}
\alias{world_heights}
\title{World Height Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1499 rows and 3 columns.}
\usage{
world_heights
}
\description{
Average male heights in 165 countries from 1810-1989, with a
smaller number of countries from 1500-1800.
}
\details{
It contains the variables:
\itemize{
\item country: The Country.
\item year: Year.
\item height_cm: Average male height in centimeters.
}

For more information, see the article: "Why are you tall while others are
short? Agricultural production and other proximate determinants of global
heights",  Joerg Baten and Matthias Blum, European Review of Economic
History 18 (2014), 144–165. Data available from
\url{http://hdl.handle.net/10622/IAEKLA}, accessed via the Clio Infra website.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(world_heights, 
       aes(x = year, 
           y = height_cm, 
           group = country)) + 
    geom_line()
}
}
\keyword{datasets}
